test: merge fuzz_test.go into main_test.go
The fuzz suite is six small targets — fits cleanly alongside the
unit tests. Single-package convention prefers fewer files when the
infrastructure isn't dominant. CI `Fuzz...` selectors and the
test.yml fuzz step are unchanged (they reference function names,
not file names).

No behavior change; coverage and fuzz seed-corpus runs identical.
